Lesson 6

+++ api/src/main/java/com/paltvlad/market/api/ProductDto.java
@@ -1,6 +1,7 @@
 package com.paltvlad.market.api;
 
 
+import java.math.BigDecimal;
 import java.util.Objects;
 
 
@@ -12,11 +13,11 @@ public class ProductDto {
 
     private String title;
 
-    private double price;
+    private BigDecimal price;
 
     private String categoryTitle;
 
-    public ProductDto(Long id, String title, double price, String categoryTitle) {
+    public ProductDto(Long id, String title, BigDecimal price, String categoryTitle) {
         this.id = id;
         this.title = title;
         this.price = price;
@@ -42,11 +43,11 @@ public class ProductDto {
         this.title = title;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
