+++ core-service/src/main/java/com/paltvlad/market/core/repositories/specifications/ProductsSpecifications.java
@@ -4,13 +4,15 @@ package com.paltvlad.market.core.repositories.specifications;
 import com.paltvlad.market.core.entities.Product;
 import org.springframework.data.jpa.domain.Specification;
 
+import java.math.BigDecimal;
+
 public class ProductsSpecifications {
 
-    public static Specification<Product> priceGreaterOrEqualsThan(Double priceMin) {
+    public static Specification<Product> priceGreaterOrEqualsThan(BigDecimal priceMin) {
         return (root, criteriaQuery, criteriaBuilder) -> criteriaBuilder.greaterThanOrEqualTo(root.get("price"), priceMin);
     }
 
-    public static Specification<Product> priceLessOrEqualsThan(Double priceMax) {
+    public static Specification<Product> priceLessOrEqualsThan(BigDecimal priceMax) {
         return (root, criteriaQuery, criteriaBuilder) -> criteriaBuilder.lessThanOrEqualTo(root.get("price"), priceMax);
     }
